+++ app/controllers/projects_controller.rb
@@ -65,8 +65,8 @@ class ProjectsController < ApplicationController
   end
 
   def crop_picture
-    @project.crop_picture(params[:crop_x].to_i, params[:crop_y].to_i,
-                         params[:crop_w].to_i, params[:crop_h].to_i,
+    @project.crop_picture(params[:crop_x].to_f, params[:crop_y].to_f,
+                         params[:crop_w].to_f, params[:crop_h].to_f,
                          params[:crop_target].to_sym)
     redirect_to @project, notice: t('project.message.imageCropped')
   end
